# src/infrastructure/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AlertOut": {
        "properties": {
          "actionTaken": {
            "anyOf": [
              {
                "enum": [
                  "guidance",
                  "reschedule",
                  "social_work_referral",
                  "lodging_coordination",
                  "transport_coordination",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actiontaken"
          },
          "category": {
            "enum": [
              "transport",
              "lodging",
              "financial",
              "schedule",
              "instructions",
              "communication",
              "health_difficulty",
              "other"
            ],
            "title": "Category",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "familyContacted": {
            "title": "Familycontacted",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "internalNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internalnote"
          },
          "milestoneId": {
            "format": "uuid",
            "title": "Milestoneid",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "patientId": {
            "format": "uuid",
            "title": "Patientid",
            "type": "string"
          },
          "resolvedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolvedat"
          },
          "risk": {
            "enum": [
              "green",
              "yellow",
              "red"
            ],
            "title": "Risk",
            "type": "string"
          },
          "status": {
            "enum": [
              "open",
              "in_progress",
              "resolved"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "patientId",
          "milestoneId",
          "category",
          "status",
          "risk",
          "familyContacted",
          "createdAt"
        ],
        "title": "AlertOut",
        "type": "object"
      },
...
      "ConfirmAttendanceRequest": {
        "properties": {
          "operationId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operationid"
          }
        },
        "title": "ConfirmAttendanceRequest",
        "type": "object"
      },
...
      "FeelingOut": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mood": {
            "enum": [
              "calm",
              "unsure",
              "worried",
              "tired"
            ],
            "title": "Mood",
            "type": "string"
          },
          "patientId": {
            "format": "uuid",
            "title": "Patientid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "patientId",
          "mood",
          "createdAt"
        ],
        "title": "FeelingOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MilestoneOut": {
        "properties": {
          "attendanceConfirmed": {
            "title": "Attendanceconfirmed",
            "type": "boolean"
          },
          "confirmationDeadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmationdeadline"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "patientId": {
            "format": "uuid",
            "title": "Patientid",
            "type": "string"
          },
          "preparation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preparation"
          },
          "scheduledAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledat"
          },
          "service": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service"
          },
          "status": {
            "enum": [
              "completed",
              "upcoming",
              "unscheduled",
              "support_needed",
              "rescheduled",
              "missed"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "enum": [
              "consultation",
              "laboratory",
              "procedure",
              "treatment",
              "follow_up"
            ],
            "title": "Type",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "patientId",
          "type",
          "title",
          "status",
          "attendanceConfirmed",
          "version"
        ],
        "title": "MilestoneOut",
        "type": "object"
      },
...
      "PatientOut": {
        "properties": {
          "age": {
            "title": "Age",
            "type": "integer"
          },
          "avatarKey": {
            "title": "Avatarkey",
            "type": "string"
          },
          "caregiverName": {
            "title": "Caregivername",
            "type": "string"
          },
          "contactPhone": {
            "title": "Contactphone",
            "type": "string"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "operationalRisk": {
            "enum": [
              "green",
              "yellow",
              "red"
            ],
            "title": "Operationalrisk",
            "type": "string"
          },
          "routeStatus": {
            "enum": [
              "on_track",
              "confirmation_needed",
              "support_needed"
            ],
            "title": "Routestatus",
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName",
          "age",
          "avatarKey",
          "caregiverName",
          "contactPhone",
          "operationalRisk",
          "routeStatus"
        ],
        "title": "PatientOut",
        "type": "object"
      },
...
      "RecordFeelingRequest": {
        "properties": {
          "mood": {
            "enum": [
              "calm",
              "unsure",
              "worried",
              "tired"
            ],
            "title": "Mood",
            "type": "string"
          },
          "operationId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operationid"
          }
        },
        "required": [
          "mood"
        ],
        "title": "RecordFeelingRequest",
        "type": "object"
      },
...
      "ReportBarrierRequest": {
        "properties": {
          "category": {
            "enum": [
              "transport",
              "lodging",
              "financial",
              "schedule",
              "instructions",
              "communication",
              "health_difficulty",
              "other"
            ],
            "title": "Category",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "operationId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operationid"
          }
        },
        "required": [
          "category"
        ],
        "title": "ReportBarrierRequest",
        "type": "object"
      },
...
      "RouteResponse": {
        "properties": {
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/AlertOut"
            },
            "title": "Alerts",
            "type": "array"
          },
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/MilestoneOut"
            },
            "title": "Milestones",
            "type": "array"
          },
          "nextMilestoneId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextmilestoneid"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientOut"
          }
        },
        "required": [
          "patient",
          "milestones",
          "alerts"
        ],
        "title": "RouteResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API del piloto conectado de Kinti (Fase 2). Datos exclusivamente sint�ticos. El sem�foro representa riesgo operativo de interrupci�n de la ruta, no gravedad cl�nica. No diagnostica, no prescribe y no realiza triaje.",
    "title": "Kinti API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/milestones/{milestone_id}/barriers": {
      "post": {
        "operationId": "report_barrier_api_v1_milestones__milestone_id__barriers_post",
        "parameters": [
          {
            "in": "path",
            "name": "milestone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Milestone Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportBarrierRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Report Barrier",
        "tags": [
          "familia"
        ]
      }
    },
    "/api/v1/milestones/{milestone_id}/confirmations": {
      "post": {
        "operationId": "confirm_attendance_api_v1_milestones__milestone_id__confirmations_post",
        "parameters": [
          {
            "in": "path",
            "name": "milestone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Milestone Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmAttendanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Confirm Attendance",
        "tags": [
          "familia"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}/feelings": {
      "post": {
        "operationId": "record_feeling_api_v1_patients__patient_id__feelings_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordFeelingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeelingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record Feeling",
        "tags": [
          "familia"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}/route": {
      "get": {
        "operationId": "get_route_api_v1_patients__patient_id__route_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Route",
        "tags": [
          "familia"
        ]
      }
    },
...
  }
}